refactor: split model kernel compatibility policy
